+++ controllers/cubes.js
@@ -12,7 +12,6 @@ const getCube = async (id) => {
 
 const getCubeWithAccessories = async (id) => {
     const cube = await Cube.findById(id).populate('accessories').lean();
-    console.log(cube)
     return cube;
 }
