+++ index.js
@@ -4,6 +4,10 @@ const mongoose = require('mongoose');
 const config = require('./config/config')[env];
 const express = require('express');
 const indexRouter = require('./routes/index');
+const authRouter = require('./routes/auth');
+const cubeRouter = require('./routes/cube');
+const accessoryRouter = require('./routes/accessory');
+const accessory = require('./models/accessory');
 const app = express();
 
 mongoose.connect(config.databaseURL,
@@ -20,6 +24,9 @@ mongoose.connect(config.databaseURL,
 
 require('./config/express')(app);
 
+app.use('/', authRouter);
+app.use('/', cubeRouter);
+app.use('/', accessoryRouter);
 app.use('/', indexRouter);
 
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`));
