+++ package.json
@@ -5,8 +5,11 @@
     "version": "1.0.0",
     "main": "index.js",
     "dependencies": {
+        "bcrypt": "^5.0.0",
+        "cookie-parser": "^1.4.5",
         "express": "^4.17.1",
         "express-handlebars": "^3.1.0",
+        "jsonwebtoken": "^8.5.1",
         "mongoose": "^5.9.18",
         "uuid": "^8.1.0"
     },
